% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecting.R
\name{connect_to_bulldrive}
\alias{connect_to_bulldrive}
\title{Quick-connect to the Bulldrive system}
\usage{
connect_to_bulldrive()
}
\description{
This function is a wrapper around the egnyter::set_token
function and presumes you have appropriate config
and environment variables set up.
}
