% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecting.R
\name{connect_to_dap}
\alias{connect_to_dap}
\title{Quick-connect to the DAP system (deprecated)}
\usage{
connect_to_dap()
}
\description{
This function is a wrapper around the DBI::dbConnect
function and presumes you have appropriate config
and environment variables set up.
}
