% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecting.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Quick-connect to an internal database using DBI}
\usage{
connect_to_database(db_name, ...)
}
\arguments{
\item{db_name}{The name of the database connection (must match config.yml)}

\item{...}{Other arguments passed to dbConnect (such as 'DB')}
}
\description{
This function is a wrapper around the DBI::dbConnect
function and presumes you have appropriate config
and environment variables set up.
}
