% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connecting.R
\name{connect_to_ods}
\alias{connect_to_ods}
\title{Quick-connect to the ODS system}
\usage{
connect_to_ods()
}
\description{
This function is a wrapper around the DBI::dbConnect
function and presumes you have appropriate config
and environment variables set up.
}
