% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_database_object}
\alias{create_database_object}
\title{Create a Bulldrive folder structure for the project}
\usage{
create_database_object(x, name, dest, type = c("view", "table"))
}
\arguments{
\item{x}{dbplyr pipe / remote tbl object}

\item{name}{name for new remote object}

\item{dest}{remote data source connection}

\item{type}{specify whether the object is a view (default or table)}
}
\description{
This function creates an appropriate project folder
structure in Bulldrive using the project name as
the top-level folder name.
}
