% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{create_project_folders}
\alias{create_project_folders}
\title{Create a Bulldrive folder structure for the project}
\usage{
create_project_folders(project_name,
  where = config::get("bulldrive")$project_folder)
}
\arguments{
\item{project_name}{The name of the project (folder)}
}
\description{
This function creates an appropriate project folder
structure in Bulldrive using the project name as
the top-level folder name.
}
