% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_database_object}
\alias{create_database_object}
\title{Create an object in a remote datasource from a tbl}
\usage{
create_database_object(x, name, dest, type = c("view", "table"))
}
\arguments{
\item{x}{dbplyr pipe / remote tbl object}

\item{name}{name for new remote object (including schema)}

\item{dest}{remote data source connection}

\item{type}{specify whether the object is a view (default or table)}
}
\description{
This function will create a remote object (view or table) in
a remote data source based on a dbplyr pipe chain.
Essentially it puts a "CREATE VIEW/TABLE <name> AS ()" wrapper
around the SQL code that is generated by the dbplyr chain.
}
